clicking on several copies switches to detail view
